requirements.txt
alembic==1.13.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cowsay==6.1
cryptography==42.0.8
cycler==0.12.1
dataclasses==0.6
devtools==0.12.2
distlib==0.3.8
distro==1.9.0
Django==5.1.2
et-xmlfile==1.1.0
executing==2.0.1
Faker==13.3.2
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Environments==0.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-profiler==1.8.1
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
fonttools==4.50.0
fsspec==2024.2.0
future==0.18.2
graphviz==0.14.2
greenlet==3.0.3
gunicorn==22.0.0
haversine==2.8.1
huggingface-hub==0.21.4
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
login==0.0.6
lxml==5.1.0
Mako==1.3.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.0.29
mysqlclient==2.2.0
mysqldump==0.0.10
netmiko==3.3.2
network==0.1
networkx==3.2.1
ntc-templates==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
paramiko==2.7.2
pillow==10.3.0
pillow_heif==0.17.0
platformdirs==4.2.1
pluggy==1.5.0
pprintpp==0.4.0
protobuf==3.19.4
pycodestyle==2.6.0
pycparser==2.21
pygeoip==0.3.2
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
pytest==8.3.2
pytest-clarity==1.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-memcached==1.62
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.32.3
rich==13.7.1
rope==0.18.0
safetensors==0.4.2
scikit-learn==1.5.0
scipy==1.12.0
scp==0.13.2
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.5.1
subprocess.run==0.0.8
sympy==1.12
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
textfsm==1.1.0
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.1
tqdm==4.66.4
transformers==4.38.2
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.2
virtualenv==20.26.0
vk==3.0
webencodings==0.5.1
Werkzeug==3.0.3
zipp==3.6.0
